design a simple ui
