--- User Interface/LearnUI/FoodTracker/FoodTracker.xcodeproj/xcuserdata/shuoyang.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -1,32 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-	<key>SchemeUserState</key>
-	<dict>
-		<key>FoodTracker.xcscheme</key>
-		<dict>
-			<key>orderHint</key>
-			<integer>0</integer>
-		</dict>
-	</dict>
-	<key>SuppressBuildableAutocreation</key>
-	<dict>
-		<key>5777D76C1DB3FB5500F95786</key>
-		<dict>
-			<key>primary</key>
-			<true/>
-		</dict>
-		<key>5777D7801DB3FB5500F95786</key>
-		<dict>
-			<key>primary</key>
-			<true/>
-		</dict>
-		<key>5777D78B1DB3FB5500F95786</key>
-		<dict>
-			<key>primary</key>
-			<true/>
-		</dict>
-	</dict>
-</dict>
-</plist>
