--- User Interface/LearnUI/FoodTracker/FoodTrackerTests/FoodTrackerTests.swift
@@ -1,36 +1,0 @@
-//
-//  FoodTrackerTests.swift
-//  FoodTrackerTests
-//
-//  Created by Shuoyang Cui on 10/16/16.
-//  Copyright © 2016 Shuoyang Cui. All rights reserved.
-//
-
-import XCTest
-@testable import FoodTracker
-
-class FoodTrackerTests: XCTestCase {
-    
-    override func setUp() {
-        super.setUp()
-        // Put setup code here. This method is called before the invocation of each test method in the class.
-    }
-    
-    override func tearDown() {
-        // Put teardown code here. This method is called after the invocation of each test method in the class.
-        super.tearDown()
-    }
-    
-    func testExample() {
-        // This is an example of a functional test case.
-        // Use XCTAssert and related functions to verify your tests produce the correct results.
-    }
-    
-    func testPerformanceExample() {
-        // This is an example of a performance test case.
-        self.measureBlock {
-            // Put the code you want to measure the time of here.
-        }
-    }
-    
-}
