--- User Interface/LearnUI/FoodTracker/FoodTrackerUITests/FoodTrackerUITests.swift
@@ -1,36 +1,0 @@
-//
-//  FoodTrackerUITests.swift
-//  FoodTrackerUITests
-//
-//  Created by Shuoyang Cui on 10/16/16.
-//  Copyright © 2016 Shuoyang Cui. All rights reserved.
-//
-
-import XCTest
-
-class FoodTrackerUITests: XCTestCase {
-        
-    override func setUp() {
-        super.setUp()
-        
-        // Put setup code here. This method is called before the invocation of each test method in the class.
-        
-        // In UI tests it is usually best to stop immediately when a failure occurs.
-        continueAfterFailure = false
-        // UI tests must launch the application that they test. Doing this in setup will make sure it happens for each test method.
-        XCUIApplication().launch()
-
-        // In UI tests it’s important to set the initial state - such as interface orientation - required for your tests before they run. The setUp method is a good place to do this.
-    }
-    
-    override func tearDown() {
-        // Put teardown code here. This method is called after the invocation of each test method in the class.
-        super.tearDown()
-    }
-    
-    func testExample() {
-        // Use recording to get started writing UI tests.
-        // Use XCTAssert and related functions to verify your tests produce the correct results.
-    }
-    
-}
