+++ User Interface/VoiceOptimizer/VoiceOptimizer.xcodeproj/xcuserdata/shuoyang.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -1,0 +1,32 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>SchemeUserState</key>
+	<dict>
+		<key>VoiceOptimizer.xcscheme</key>
+		<dict>
+			<key>orderHint</key>
+			<integer>0</integer>
+		</dict>
+	</dict>
+	<key>SuppressBuildableAutocreation</key>
+	<dict>
+		<key>57EE74321DBDACEF00B2A176</key>
+		<dict>
+			<key>primary</key>
+			<true/>
+		</dict>
+		<key>57EE74461DBDACEF00B2A176</key>
+		<dict>
+			<key>primary</key>
+			<true/>
+		</dict>
+		<key>57EE74511DBDACEF00B2A176</key>
+		<dict>
+			<key>primary</key>
+			<true/>
+		</dict>
+	</dict>
+</dict>
+</plist>
