+++ User Interface/VoiceOptimizer/VoiceOptimizer/Assets.xcassets/Contents.json
