+++ User Interface/VoiceOptimizer/VoiceOptimizer/Assets.xcassets/Speak.imageset/Contents.json
@@ -1,0 +1,13 @@
+{
+  "images" : [
+    {
+      "idiom" : "universal",
+      "filename" : "Voice Recognition Scan Filled-50.png",
+      "scale" : "1x"
+    }
+  ],
+  "info" : {
+    "version" : 1,
+    "author" : "xcode"
+  }
+}
