+++ voiceOptimizer/voiceOptimizer.xcodeproj/xcuserdata/shuoyang.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -1,0 +1,32 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>SchemeUserState</key>
+	<dict>
+		<key>voiceOptimizer.xcscheme</key>
+		<dict>
+			<key>orderHint</key>
+			<integer>0</integer>
+		</dict>
+	</dict>
+	<key>SuppressBuildableAutocreation</key>
+	<dict>
+		<key>0F1909081DB40984003AD086</key>
+		<dict>
+			<key>primary</key>
+			<true/>
+		</dict>
+		<key>0F19091C1DB40985003AD086</key>
+		<dict>
+			<key>primary</key>
+			<true/>
+		</dict>
+		<key>0F1909271DB40985003AD086</key>
+		<dict>
+			<key>primary</key>
+			<true/>
+		</dict>
+	</dict>
+</dict>
+</plist>
